refactor delta module
