+++ encodings/fastlanes/src/delta/mod.rs
@@ -1,344 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-use std::fmt::Debug;
-use std::hash::Hash;
+mod array;
+pub use array::DeltaArray;
+pub use array::delta_compress::delta_compress;
 
-pub use compress::*;
-use fastlanes::FastLanes;
-use prost::Message;
-use vortex_array::arrays::PrimitiveArray;
-use vortex_array::serde::ArrayChildren;
-use vortex_array::stats::{ArrayStats, StatsSetRef};
-use vortex_array::validity::Validity;
-use vortex_array::vtable::{
-    ArrayVTable, CanonicalVTable, NotSupported, VTable, ValidityChildSliceHelper,
-    ValidityVTableFromChildSliceHelper, VisitorVTable,
-};
-use vortex_array::{
-    Array, ArrayBufferVisitor, ArrayChildVisitor, ArrayEq, ArrayHash, ArrayRef, Canonical,
-    EncodingId, EncodingRef, IntoArray, Precision, ProstMetadata, vtable,
-};
-use vortex_buffer::{Buffer, ByteBuffer};
-use vortex_dtype::{DType, NativePType, PType, match_each_unsigned_integer_ptype};
-use vortex_error::{VortexExpect as _, VortexResult, vortex_bail, vortex_err};
+pub mod compute;
 
-mod compress;
-mod compute;
-mod ops;
-
-vtable!(Delta);
-
-#[derive(Clone, prost::Message)]
-#[repr(C)]
-pub struct DeltaMetadata {
-    #[prost(uint64, tag = "1")]
-    deltas_len: u64,
-    #[prost(uint32, tag = "2")]
-    offset: u32, // must be <1024
-}
-
-impl VTable for DeltaVTable {
-    type Array = DeltaArray;
-    type Encoding = DeltaEncoding;
-    type Metadata = ProstMetadata<DeltaMetadata>;
-
-    type ArrayVTable = Self;
-    type CanonicalVTable = Self;
-    type OperationsVTable = Self;
-    type ValidityVTable = ValidityVTableFromChildSliceHelper;
-    type VisitorVTable = Self;
-    type ComputeVTable = NotSupported;
-    type EncodeVTable = NotSupported;
-    type OperatorVTable = NotSupported;
-
-    fn id(_encoding: &Self::Encoding) -> EncodingId {
-        EncodingId::new_ref("fastlanes.delta")
-    }
-
-    fn encoding(_array: &Self::Array) -> EncodingRef {
-        EncodingRef::new_ref(DeltaEncoding.as_ref())
-    }
-
-    fn metadata(array: &DeltaArray) -> VortexResult<Self::Metadata> {
-        Ok(ProstMetadata(DeltaMetadata {
-            deltas_len: array.deltas().len() as u64,
-            offset: array.offset() as u32,
-        }))
-    }
-
-    fn serialize(metadata: Self::Metadata) -> VortexResult<Option<Vec<u8>>> {
-        Ok(Some(metadata.0.encode_to_vec()))
-    }
-
-    fn deserialize(buffer: &[u8]) -> VortexResult<Self::Metadata> {
-        Ok(ProstMetadata(DeltaMetadata::decode(buffer)?))
-    }
-
-    fn build(
-        _encoding: &DeltaEncoding,
-        dtype: &DType,
-        len: usize,
-        metadata: &Self::Metadata,
-        _buffers: &[ByteBuffer],
-        children: &dyn ArrayChildren,
-    ) -> VortexResult<DeltaArray> {
-        assert_eq!(children.len(), 2);
-        let ptype = PType::try_from(dtype)?;
-        let lanes = match_each_unsigned_integer_ptype!(ptype, |T| { <T as FastLanes>::LANES });
-
-        // Compute the length of the bases array
-        let deltas_len = usize::try_from(metadata.0.deltas_len)
-            .map_err(|_| vortex_err!("deltas_len {} overflowed usize", metadata.0.deltas_len))?;
-        let num_chunks = deltas_len / 1024;
-        let remainder_base_size = if deltas_len % 1024 > 0 { 1 } else { 0 };
-        let bases_len = num_chunks * lanes + remainder_base_size;
-
-        let bases = children.get(0, dtype, bases_len)?;
-        let deltas = children.get(1, dtype, deltas_len)?;
-
-        DeltaArray::try_new(bases, deltas, metadata.0.offset as usize, len)
-    }
-}
-
-/// A FastLanes-style delta-encoded array of primitive values.
-///
-/// A [`DeltaArray`] comprises a sequence of _chunks_ each representing 1,024 delta-encoded values,
-/// except the last chunk which may represent from one to 1,024 values.
-///
-/// # Examples
-///
-/// ```
-/// use vortex_fastlanes::DeltaArray;
-/// let array = DeltaArray::try_from_vec(vec![1_u32, 2, 3, 5, 10, 11]).unwrap();
-/// ```
-///
-/// # Details
-///
-/// To facilitate slicing, this array accepts an `offset` and `logical_len`. The offset must be
-/// strictly less than 1,024 and the sum of `offset` and `logical_len` must not exceed the length of
-/// the `deltas` array. These values permit logical slicing without modifying any chunk containing a
-/// kept value. In particular, we may defer decompresison until the array is canonicalized or
-/// indexed. The `offset` is a physical offset into the first chunk, which necessarily contains
-/// 1,024 values. The `logical_len` is the number of logical values following the `offset`, which
-/// may be less than the number of physically stored values.
-///
-/// Each chunk is stored as a vector of bases and a vector of deltas. If the chunk physically
-/// contains 1,024 values, then there are as many bases as there are _lanes_ of this type in a
-/// 1024-bit register. For example, for 64-bit values, there are 16 bases because there are 16
-/// _lanes_. Each lane is a [delta-encoding](https://en.wikipedia.org/wiki/Delta_encoding) `1024 /
-/// bit_width` long vector of values. The deltas are stored in the
-/// [FastLanes](https://www.vldb.org/pvldb/vol16/p2132-afroozeh.pdf) order which splits the 1,024
-/// values into one contiguous sub-sequence per-lane, thus permitting delta encoding.
-///
-/// If the chunk physically has fewer than 1,024 values, then it is stored as a traditional,
-/// non-SIMD-amenable, delta-encoded vector.
-///
-/// Note the validity is stored in the deltas array.
-#[derive(Clone, Debug)]
-pub struct DeltaArray {
-    offset: usize,
-    len: usize,
-    dtype: DType,
-    bases: ArrayRef,
-    deltas: ArrayRef,
-    stats_set: ArrayStats,
-}
-
-#[derive(Clone, Debug)]
-pub struct DeltaEncoding;
-
-impl DeltaArray {
-    // TODO(ngates): remove constructing from vec
-    pub fn try_from_vec<T: NativePType>(vec: Vec<T>) -> VortexResult<Self> {
-        Self::try_from_primitive_array(&PrimitiveArray::new(
-            Buffer::copy_from(vec),
-            Validity::NonNullable,
-        ))
-    }
-
-    pub fn try_from_primitive_array(array: &PrimitiveArray) -> VortexResult<Self> {
-        let (bases, deltas) = delta_compress(array)?;
-
-        Self::try_from_delta_compress_parts(bases.into_array(), deltas.into_array())
-    }
-
-    /// Create a [`DeltaArray`] from the given `bases` and `deltas` arrays.
-    /// Note the `deltas` might be nullable
-    pub fn try_from_delta_compress_parts(bases: ArrayRef, deltas: ArrayRef) -> VortexResult<Self> {
-        let logical_len = deltas.len();
-        Self::try_new(bases, deltas, 0, logical_len)
-    }
-
-    pub fn try_new(
-        bases: ArrayRef,
-        deltas: ArrayRef,
-        offset: usize,
-        logical_len: usize,
-    ) -> VortexResult<Self> {
-        if offset >= 1024 {
-            vortex_bail!("offset must be less than 1024: {}", offset);
-        }
-        if offset + logical_len > deltas.len() {
-            vortex_bail!(
-                "offset + logical_len, {} + {}, must be less than or equal to the size of deltas: {}",
-                offset,
-                logical_len,
-                deltas.len()
-            )
-        }
-        if !bases.dtype().eq_ignore_nullability(deltas.dtype()) {
-            vortex_bail!(
-                "DeltaArray: bases and deltas must have the same dtype, got {:?} and {:?}",
-                bases.dtype(),
-                deltas.dtype()
-            );
-        }
-        let DType::Primitive(ptype, _) = bases.dtype().clone() else {
-            vortex_bail!(
-                "DeltaArray: dtype must be an integer, got {}",
-                bases.dtype()
-            );
-        };
-
-        if !ptype.is_int() {
-            vortex_bail!("DeltaArray: ptype must be an integer, got {}", ptype);
-        }
-
-        let lanes = lane_count(ptype);
-
-        if (deltas.len() % 1024 == 0) != (bases.len() % lanes == 0) {
-            vortex_bail!(
-                "deltas length ({}) is a multiple of 1024 iff bases length ({}) is a multiple of LANES ({})",
-                deltas.len(),
-                bases.len(),
-                lanes,
-            );
-        }
-
-        // SAFETY: validation done above
-        Ok(unsafe { Self::new_unchecked(bases, deltas, offset, logical_len) })
-    }
-
-    pub(crate) unsafe fn new_unchecked(
-        bases: ArrayRef,
-        deltas: ArrayRef,
-        offset: usize,
-        logical_len: usize,
-    ) -> Self {
-        Self {
-            offset,
-            len: logical_len,
-            dtype: bases.dtype().with_nullability(deltas.dtype().nullability()),
-            bases,
-            deltas,
-            stats_set: Default::default(),
-        }
-    }
-
-    #[inline]
-    pub fn bases(&self) -> &ArrayRef {
-        &self.bases
-    }
-
-    #[inline]
-    pub fn deltas(&self) -> &ArrayRef {
-        &self.deltas
-    }
-
-    #[inline]
-    fn lanes(&self) -> usize {
-        let ptype =
-            PType::try_from(self.dtype()).vortex_expect("DeltaArray DType must be primitive");
-        lane_count(ptype)
-    }
-
-    #[inline]
-    /// The logical offset into the first chunk of [`Self::deltas`].
-    pub fn offset(&self) -> usize {
-        self.offset
-    }
-
-    fn bases_len(&self) -> usize {
-        self.bases.len()
-    }
-
-    fn deltas_len(&self) -> usize {
-        self.deltas.len()
-    }
-}
-
-pub(crate) fn lane_count(ptype: PType) -> usize {
-    match_each_unsigned_integer_ptype!(ptype, |T| { T::LANES })
-}
-
-impl ValidityChildSliceHelper for DeltaArray {
-    fn unsliced_child_and_slice(&self) -> (&ArrayRef, usize, usize) {
-        let (start, len) = (self.offset(), self.len());
-        (self.deltas(), start, start + len)
-    }
-}
-
-impl ArrayVTable<DeltaVTable> for DeltaVTable {
-    fn len(array: &DeltaArray) -> usize {
-        array.len
-    }
-
-    fn dtype(array: &DeltaArray) -> &DType {
-        &array.dtype
-    }
-
-    fn stats(array: &DeltaArray) -> StatsSetRef<'_> {
-        array.stats_set.to_ref(array.as_ref())
-    }
-
-    fn array_hash<H: std::hash::Hasher>(array: &DeltaArray, state: &mut H, precision: Precision) {
-        array.offset.hash(state);
-        array.len.hash(state);
-        array.dtype.hash(state);
-        array.bases.array_hash(state, precision);
-        array.deltas.array_hash(state, precision);
-    }
-
-    fn array_eq(array: &DeltaArray, other: &DeltaArray, precision: Precision) -> bool {
-        array.offset == other.offset
-            && array.len == other.len
-            && array.dtype == other.dtype
-            && array.bases.array_eq(&other.bases, precision)
-            && array.deltas.array_eq(&other.deltas, precision)
-    }
-}
-
-impl CanonicalVTable<DeltaVTable> for DeltaVTable {
-    fn canonicalize(array: &DeltaArray) -> Canonical {
-        Canonical::Primitive(delta_decompress(array))
-    }
-}
-
-impl VisitorVTable<DeltaVTable> for DeltaVTable {
-    fn visit_buffers(_array: &DeltaArray, _visitor: &mut dyn ArrayBufferVisitor) {}
-
-    fn visit_children(array: &DeltaArray, visitor: &mut dyn ArrayChildVisitor) {
-        visitor.visit_child("bases", array.bases());
-        visitor.visit_child("deltas", array.deltas());
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use vortex_array::test_harness::check_metadata;
-
-    use super::{DeltaMetadata, ProstMetadata};
-
-    #[cfg_attr(miri, ignore)]
-    #[test]
-    fn test_delta_metadata() {
-        check_metadata(
-            "delta.metadata",
-            ProstMetadata(DeltaMetadata {
-                offset: u32::MAX,
-                deltas_len: u64::MAX,
-            }),
-        );
-    }
-}
+mod vtable;
+pub use vtable::{DeltaEncoding, DeltaVTable};
